Refactor fixture projects
