+++ acceptance-test/README.md
@@ -9,3 +9,5 @@ This contains example projects to verify behavior of the plugin.
 - [Using custom template for code generation](custom-template/build.gradle)
 - [Using custom generator class for code generation](custom-class/build.gradle)
 - [Publishing template to external repository](externalize-template/build.gradle)
+- [Publishing generator class to external repository](externalize-template/build.gradle)
+- [Handling multiple sources](multiple-sources/build.gradle)
