--- acceptance-test/html-generator/build.gradle
@@ -1,21 +1,0 @@
-plugins {
-    id 'org.hidetake.swagger.generator'
-}
-
-repositories {
-    jcenter()
-}
-
-dependencies {
-    swaggerCodegen 'io.swagger:swagger-codegen-cli:2.3.1'
-}
-
-swaggerSources {
-    petstore {
-        code {
-            language = 'html2'
-            inputFile = file("$buildDir/petstore.yaml")
-            outputDir = file("$buildDir/swagger-html")
-        }
-    }
-}
