--- acceptance-test/src/test/groovy/HtmlGeneratorSpec.groovy
@@ -1,45 +1,0 @@
-import org.gradle.testkit.runner.GradleRunner
-import org.gradle.testkit.runner.TaskOutcome
-import spock.lang.Specification
-
-import static Fixture.cleanBuildDir
-import static Fixture.placePetstoreYaml
-
-class HtmlGeneratorSpec extends Specification {
-
-    GradleRunner runner
-
-    def setup() {
-        runner = GradleRunner.create()
-            .withProjectDir(new File('html-generator'))
-            .withPluginClasspath()
-            .forwardOutput()
-        cleanBuildDir(runner)
-    }
-
-    def 'generateSwaggerCode task should generate a document'() {
-        given:
-        placePetstoreYaml(runner, Fixture.PetstoreYaml.valid)
-        runner.withArguments('--stacktrace', 'generateSwaggerCode')
-
-        when:
-        def result = runner.build()
-
-        then:
-        result.task(':generateSwaggerCode').outcome == TaskOutcome.NO_SOURCE
-        result.task(':generateSwaggerCodePetstore').outcome == TaskOutcome.SUCCESS
-        new File(runner.projectDir, 'build/swagger-html/index.html').exists()
-    }
-
-    def 'generateSwaggerCodePetstoreHelp task should show help'() {
-        given:
-        runner.withArguments('--stacktrace', 'generateSwaggerCodePetstoreHelp')
-
-        when:
-        def result = runner.build()
-
-        then:
-        result.output.contains('CONFIG OPTIONS')
-    }
-
-}
